0665: Actually working solution
Instead of being overly fancy and wanting to complicate the code for the sake of only iterating through the list once,
here we just check if each number is equal to or greater than the last, and if this is not the case at any point we
just drop both the last number (which might be too high to fit in the list) or the current number (which might be
too low to fit in the list) and run the modified list through the same algorithm. There is no need to modify the
number ourselves since all we need to know is whether one number can be changed to somethign (anything) and create
an increasing list.

Definitely worth considering better solutions here, however...